% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{as_daily}
\alias{as_daily}
\title{Cast Table to Daily}
\usage{
as_daily(data)
}
\arguments{
\item{data}{A data frame or tibble to be used as the underlying data.}
}
\value{
An object of class \code{ghcn_daily}.
}
\description{
Cast Table to Daily
}
\examples{
\dontrun{
df <- read.csv(...)
df <- as_daily(df)
}
}
