% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CA003076680}
\alias{CA003076680}
\title{Maximum Temperature for Station CA003076680}
\format{
\subsection{\code{CA003076680}}{

A 'ghcn-daily' object, i.e. table 7,574 x 8:
\describe{
\item{date}{Date of measurment}
\item{station}{Station name, i.e. 'CA003076680'}
\item{tmax}{Maximum temperature}
\item{tmin}{Minimum temperature}
\item{prcp}{Total precipitation}
\item{*_flag}{Flags for the measurments}
}
}
}
\source{
\url{https://www.countrycallingcodes.com/iso-country-codes/europe-codes.php}
}
\usage{
CA003076680
}
\description{
Maximum Temperature for Station CA003076680
}
\keyword{datasets}
