% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{Countries ISO Codes}
\format{
\subsection{\code{europe_codes}}{

A table 253 x 2:
\describe{
\item{name}{Country name}
\item{iso3}{3 letter ISO country code}
}
}
}
\usage{
country_codes
}
\description{
Countries ISO Codes
}
\keyword{datasets}
