% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{daily}
\alias{daily}
\title{Download Daily Summaries}
\usage{
daily(station_id, start_date, end_date)
}
\arguments{
\item{station_id}{Character, station id(s).}

\item{start_date}{Character, start date.}

\item{end_date}{Character, end date.}
}
\value{
A tibble with the daily timeseries at the stations.
}
\description{
Download Daily Summaries
}
\details{
\emph{station_id} can be a vector with multiple stations.
Dates should be given in \code{YYYY-mm-dd} format.
}
