% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.add_variables}
\alias{.add_variables}
\title{Add Columns to Handle Summarize}
\usage{
.add_variables(x)
}
\arguments{
\item{x}{Object of class \code{ghcn_daily}. See \code{\link[=daily]{daily()}} for details.}
}
\value{
Table with number of days in the months.
}
\description{
Add Columns to Handle Summarize
}
