% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.api_error}
\alias{.api_error}
\title{Handles API Errors}
\usage{
.api_error(resp)
}
\arguments{
\item{resp}{Object of class \code{httr2_response}.}
}
\value{
NULL, called for side effects.
}
\description{
Handles API Errors
}
