% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{.check_flags}
\alias{.check_flags}
\title{Check Flags Columns}
\usage{
.check_flags(x)
}
\arguments{
\item{x}{Object of class \code{ghcn_daily}. See \code{\link[=daily]{daily()}} for details.}
}
\value{
NULL, called for side effects
}
\description{
Check Flags Columns
}
