% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{.drop_flags}
\alias{.drop_flags}
\title{Drop Flags Columns}
\usage{
.drop_flags(x)
}
\arguments{
\item{x}{Object of class \code{ghcn_daily}. See \code{\link[=daily]{daily()}} for details.}
}
\value{
The original objects without flags column.
}
\description{
Drop Flags Columns
}
