% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{.extract_flag}
\alias{.extract_flag}
\title{Extract GHCNd Flags}
\usage{
.extract_flag(x)
}
\arguments{
\item{x}{Character, vector of the flag as returned by the GHCNd API call.}
}
\value{
Character of the flag.
}
\description{
Extract GHCNd Flags
}
\details{
\url{https://www.ncei.noaa.gov/products/land-based-station/global-historical-climatology-network-daily}
}
