% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{.flags}
\alias{.flags}
\title{GHCNd Flags}
\usage{
.flags()
}
\value{
Table with flags.
}
\description{
GHCNd Flags
}
\details{
\url{https://www.ncei.noaa.gov/products/land-based-station/global-historical-climatology-network-daily}
}
