% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{.s3_annual}
\alias{.s3_annual}
\title{Annual Class Constructor}
\usage{
.s3_annual(data = tibble::tibble())
}
\arguments{
\item{data}{A data frame or tibble to be used as the underlying data.}
}
\value{
An object of class \code{ghcn_annual}.
}
\description{
Annual Class Constructor
}
\details{
Creates a new object of class \code{ghcn_annual}.
}
