% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{.s3_anomaly}
\alias{.s3_anomaly}
\title{Anomaly Constructor}
\usage{
.s3_anomaly(data = tibble::tibble())
}
\arguments{
\item{data}{A data frame or tibble to be used as the underlying data.}
}
\value{
An object of class \code{ghcn_anomaly}.
}
\description{
Anomaly Constructor
}
\details{
Creates a new object of class \code{ghcn_anomaly}.
}
