% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{.s3_monthly}
\alias{.s3_monthly}
\title{Daily Class Constructor}
\usage{
.s3_monthly(data = tibble::tibble())
}
\arguments{
\item{data}{A data frame or tibble to be used as the underlying data.}
}
\value{
An object of class \code{monthly}.
}
\description{
Daily Class Constructor
}
\details{
Creates a new object of class \code{monthly}.
}
