% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_classes.R
\name{.s3_quarterly}
\alias{.s3_quarterly}
\title{Annual Quarter Constructor}
\usage{
.s3_quarterly(data = tibble::tibble())
}
\arguments{
\item{data}{A data frame or tibble to be used as the underlying data.}
}
\value{
An object of class \code{ghcn_quarterly}.
}
\description{
Annual Quarter Constructor
}
\details{
Creates a new object of class \code{ghcn_quarterly}.
}
