% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{download_inventory}
\alias{download_inventory}
\title{Download GHCNd Inventory File}
\usage{
download_inventory(filename)
}
\arguments{
\item{filename}{Character of the filename of the inventory, if already downloaded.}
}
\value{
Character, the location of the file where the inventory has been saved.
}
\description{
Download GHCNd Inventory File
}
\details{
Download the inventory from <"https://www.ncei.noaa.gov/pub/data/ghcn/daily/ghcnd-inventory.txt">.
}
