% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation.R
\name{elevation_stations}
\alias{elevation_stations}
\title{Get GHCNd Station Elevation}
\usage{
elevation_stations()
}
\value{
The table with the elevation of GHCNd stations.
}
\description{
Get GHCNd Station Elevation
}
\examples{
\dontrun{
el <- elevation_stations()
}
}
