% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{europe_codes}
\alias{europe_codes}
\title{Countries ISO Codes}
\format{
\subsection{\code{europe_codes}}{

A table 50 x 3:
\describe{
\item{name}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
}
}
}
\source{
\url{https://www.countrycallingcodes.com/iso-country-codes/europe-codes.php}
}
\usage{
europe_codes
}
\description{
Countries ISO Codes
}
\keyword{datasets}
