% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{filter_stations}
\alias{filter_stations}
\title{Spatial Filtering of Stations}
\usage{
filter_stations(stations, roi)
}
\arguments{
\item{stations, }{the table with station data. See \code{\link[=stations]{stations()}}.}

\item{roi, }{the geometry of the region of interest. See \code{\link[=get_country]{get_country()}}.}
}
\value{
Table with filtered stations.
}
\description{
Spatial Filtering of Stations
}
\examples{
\dontrun{
inventory <- stations()
roi <- get_country("ITA")
s <- filter_stations(inventory, roi)
}
}
