% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-country-shapefile.R
\name{get_countries}
\alias{get_countries}
\title{Download multiple countries' shapefiles from geoBoundaries}
\usage{
get_countries(countries_code, simplified = TRUE)
}
\arguments{
\item{countries_code}{Vector of three letter ISO code.}

\item{simplified}{Logical.}
}
\value{
A shapefile.
}
\description{
Download multiple countries' shapefiles from geoBoundaries
}
\details{
\url{https://github.com/wmgeolab/geoBoundaries}.
}
