% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-ghcn.R
\name{get_daily_weather}
\alias{get_daily_weather}
\title{Download data from GHCN}
\usage{
get_daily_weather(country_code, variable, start_date, end_date)
}
\arguments{
\item{country_code}{3 letter ISO-3 country code.}

\item{variable}{either TMAX, TMIN, or TAVG.}

\item{start_date}{Start date as character (\%YYYY-\%mm-\%dd).}

\item{end_date}{Start date as character (\%YYYY-\%mm-\%dd).}
}
\value{
A tibble with the stations within the \code{roi}.
}
\description{
Download data from GHCN
}
\details{
Dates should be given in \code{YYYY-mm-dd} format.
A list of ISO-3 for Europe is in GHCNr::europe_codes.
}
