% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{monthly}
\alias{monthly}
\title{Calculate Monthly Summaries}
\usage{
monthly(x)
}
\arguments{
\item{x}{Object of class \code{ghcn_daily}. See \code{\link[=daily]{daily()}} for details.}
}
\value{
A tibble with the monthly timeseries at the stations.
}
\description{
Calculate Monthly Summaries
}
\details{
\emph{x} is the table returned from \code{daily()} or
\code{remove_flagged()} or any subset of them.
}
\examples{
monthly(CA003076680)
}
