% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.ghcn_annual}
\alias{plot.ghcn_annual}
\title{Plot GHCN Timeseries}
\usage{
\method{plot}{ghcn_annual}(x, variable, ...)
}
\arguments{
\item{x}{Object of class ghcn_annual.}

\item{variable}{Name of the variable to plot.}

\item{...}{additional arguments to be passed to \code{plot()}.}
}
\description{
Plot GHCN Timeseries
}
\examples{
plot(annual(CA003076680), "tmax")
}
