% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.ghcn_anomaly}
\alias{plot.ghcn_anomaly}
\title{Plot GHCN Timeseries}
\usage{
\method{plot}{ghcn_anomaly}(x, cutoff, ...)
}
\arguments{
\item{x}{Object of class \code{ghcn_daily}. See \code{\link[=daily]{daily()}} for details.}

\item{cutoff}{Numeric, the year used as cutoff for baseline.}

\item{...}{additional arguments to be passed to \code{\link[stats:interaction.plot]{stats::interaction.plot()}}.}
}
\value{
NULL, called for side effects.
}
\description{
Plot GHCN Timeseries
}
\examples{
plot(anomaly(remove_flagged(CA003076680), 2015))
}
