% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.ghcn_quarterly}
\alias{plot.ghcn_quarterly}
\title{Plot GHCNd Timeseries}
\usage{
\method{plot}{ghcn_quarterly}(x, variable, ...)
}
\arguments{
\item{x}{Object of class \code{ghcn_quarterly}. See \code{\link[=daily]{daily()}} for details.}

\item{variable}{Name of the variable to plot.}

\item{...}{additional arguments to be passed to \code{\link[stats:interaction.plot]{stats::interaction.plot()}}.}
}
\value{
NULL, called for side effects.
}
\description{
Plot GHCNd Timeseries
}
\examples{
plot(monthly(CA003076680), "tmax")
}
