% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{remove_flagged}
\alias{remove_flagged}
\title{Remove Flagged Recrods}
\usage{
remove_flagged(x)
}
\arguments{
\item{x}{Object of class \code{daily}.}
}
\value{
A tibble with the stations within the \code{roi}.
}
\description{
Remove Flagged Recrods
}
\details{
dates should be given in \code{YYYY-mm-dd} format.
}
