% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations_elevation}
\alias{stations_elevation}
\title{Elevation of the GHCNd Stations}
\format{
\subsection{\code{europe_codes}}{

A table 128,024 x 2:
\describe{
\item{station}{Station name}
\item{elevation}{Elevation in meters}
}
}
}
\source{
\url{https://www.ncei.noaa.gov/pub/data/ghcn/daily/ghcnd-stations.csv}
}
\usage{
stations_elevation
}
\description{
Elevation of the GHCNd Stations
}
\keyword{datasets}
