% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USC00010655}
\alias{USC00010655}
\title{Daily data for Station USC00010655}
\format{
\subsection{\code{USC00010655}}{

A 'ghcn-daily' object, i.e. table 7,809 x 8:
\describe{
\item{date}{Date of measurment}
\item{station}{Station name, i.e. 'USC00010655'}
\item{tmax}{Maximum temperature}
\item{tmin}{Minimum temperature}
\item{prcp}{Total precipitation}
\item{*_flag}{Flags for the measurments}
}
}
}
\source{
\url{https://www.countrycallingcodes.com/iso-country-codes/europe-codes.php}
}
\usage{
USC00010655
}
\description{
Daily data for Station USC00010655
}
\keyword{datasets}
