% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.R
\name{annual}
\alias{annual}
\title{Calculate Annual Timseries}
\usage{
annual(x)
}
\arguments{
\item{x}{Object of class \code{ghcn_daily}. See \code{\link[=daily]{daily()}} for details.}
}
\value{
A tibble with the annual timeseries at the stations.
}
\description{
\code{annual()} aggregates the daily timeseries into an annual one.
Aggregation is done differently for TMIN, TMAX, and PRCP.
}
\details{
Aggregation is done as:
\describe{
\item{TMAX}{Maximum temperature recorded in the year}
\item{TMIN}{Minimum temperature recorded in the year}
\item{PRCP}{Total (cumulative) precipitation amount in the year}
}
}
\examples{
annual(CA003076680)
}
