% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.R
\name{annual}
\alias{annual}
\title{Calculate Annual Summaries}
\usage{
annual(x)
}
\arguments{
\item{x}{Object of class \code{ghcn_daily}. See \code{\link[=daily]{daily()}} for details.}
}
\value{
A tibble with the annual timeseries at the stations.
}
\description{
Calculate Annual Summaries
}
\examples{
annual(CA003076680)
}
