% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coverage}
\alias{coverage}
\title{Calculate Coverage of Daily Summaries}
\usage{
coverage(x)
}
\arguments{
\item{x}{Object of class \code{daily}.}
}
\value{
A tibble with the stations within the \code{roi}.
}
\description{
Calculate Coverage of Daily Summaries
}
\details{
dates should be given in \code{YYYY-mm-dd} format.
}
