% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Calculate Coverage of Daily Summaries}
\usage{
coverage(x, graph)
}
\arguments{
\item{x}{Object of class \code{ghcn_daily}.}

\item{graph}{Logical, if to show a graph of annual coverage.}
}
\value{
A table with coverage.
}
\description{
Calculate Coverage of Daily Summaries
}
\details{
This function calculates the temporal coverage of stations.
It returns a table with:
\itemize{
\item mothly_coverage The proportion of the days with records in the month
\item annual_coverage The proportion of the days with records in the year
\item annual_coverage The proportion of the years with records in the reference period
}
Important: that 'annual_coverage = 1' does not mean that all years have 'annual_coverage = 1',
but rather that all years have at least one record.
}
\examples{
cleaned <- remove_flagged(CA003076680)
cover <- coverage(cleaned)
cover[cover$month == 1, ]
}
