% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{daily}
\alias{daily}
\title{Download daily summaries}
\usage{
daily(station_id, start_date, end_date)
}
\arguments{
\item{station_id}{Character vector of station id(s).}

\item{start_date}{Start date as character.}

\item{end_date}{Start date as character.}
}
\value{
A tibble with the stations within the \code{roi}.
}
\description{
Download daily summaries
}
\details{
dates should be given in \code{YYYY-mm-dd} format.
}
