% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{.daily_request}
\alias{.daily_request}
\title{Request Daily Summaries}
\usage{
.daily_request(url)
}
\arguments{
\item{url}{Character, URL of the request.}
}
\value{
Body of the JSON request.
}
\description{
Request Daily Summaries
}
