% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{.daily_request}
\alias{.daily_request}
\title{Create Request URL for Daily Summaries}
\usage{
.daily_request(station_id, start_date, end_date, variables, include_flags)
}
\arguments{
\item{station_id}{Character, station id(s).}

\item{start_date}{Character, start date.}

\item{end_date}{Character, end date.}

\item{variables}{Character, vector of the variables to include.}

\item{include_flags}{Logical, if to include flagged records.}
}
\value{
Character string with the API URL.
}
\description{
Create Request URL for Daily Summaries
}
\details{
\emph{station_id} can be a vector with multiple stations.
Dates should be given in \code{YYYY-mm-dd} format.
}
