% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.flags}
\alias{.flags}
\title{GHCNd Flags}
\usage{
.flags(strict)
}
\arguments{
\item{strict}{Logical, if to include all flags (TRUE) or not (FALSE).}
}
\value{
Table with flags.
}
\description{
GHCNd Flags
}
\details{
\url{https://doi.org/10.1175/2010JAMC2375.1}
}
