% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.missing_variables}
\alias{.missing_variables}
\title{Check Which Variables Are Absent}
\usage{
.missing_variables(x)
}
\arguments{
\item{x}{Object of class \code{ghcn_daily}.}
}
\value{
Character vector
}
\description{
Check Which Variables Are Absent
}
