% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{europe_codes}
\alias{europe_codes}
\title{World Health Organization TB data}
\format{
\subsection{\code{countries}}{

A table of 50 x 3 with columns:
\describe{
\item{name}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
}
}
}
\source{
\url{https://www.countrycallingcodes.com/iso-country-codes/europe-codes.php}
}
\usage{
europe_codes
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
