% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly.R
\name{monthly}
\alias{monthly}
\title{Download Monthly Summaries}
\usage{
monthly(station_id, start_date, end_date, only_complete = FALSE)
}
\arguments{
\item{station_id}{Character, station id(s).}

\item{start_date}{Character, start date.}

\item{end_date}{Character, end date.}

\item{only_complete}{Logical, if to retain only months with complete records.}
}
\value{
A tibble with the monthly timeseries at the stations.
}
\description{
Download Monthly Summaries
}
\details{
\emph{station_id} can be a vector with multiple stations.
Dates should be given in \code{YYYY-mm-dd} format.
Monthly summaries are derived from the daily timeseries.
February is always assumed to have 28 days when calculating coverage;0
in leap years, values of coverage > 1 are set to 1.
}
