% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{monthly_coverage}
\alias{monthly_coverage}
\title{Calculate Monthly Coverage}
\usage{
monthly_coverage(x)
}
\arguments{
\item{x}{Object of class \code{ghcn_daily}. See \code{\link[=daily]{daily()}} for details.}
}
\value{
A table with mothly coverages.
}
\description{
Calculate Monthly Coverage
}
\details{
Calculates the monthly coverage of one station.
}
\examples{
cleaned <- remove_flagged(CA003076680)
cover <- monthly_coverage(cleaned)
cover[cover$year == 2020, ]
}
