% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{period_coverage}
\alias{period_coverage}
\title{Calculate Period Coverage}
\usage{
period_coverage(x)
}
\arguments{
\item{x}{Object of class \code{ghcn_daily}.}
}
\value{
A table with coverage across the period.
}
\description{
Calculate Period Coverage
}
\details{
Calculates the period coverage of one station.
}
\examples{
cleaned <- remove_flagged(CA003076680)
cover <- period_coverage(cleaned)
cover
}
