% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.ghcn_daily}
\alias{plot.ghcn_daily}
\title{Plot GHCN Timeseries}
\usage{
\method{plot}{ghcn_daily}(x, variable, ...)
}
\arguments{
\item{x}{Object of class \code{ghcn_daily}. See \code{\link[=daily]{daily()}} for details.}

\item{variable}{Name of the variable to plot.}

\item{...}{additional arguments to be passed to \code{plot()}.}
}
\value{
NULL, called for side effects.
}
\description{
Plot GHCN Timeseries
}
\examples{
plot(CA003076680, "tmax")
}
