% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{remove_flagged}
\alias{remove_flagged}
\title{Remove Flagged Recrods}
\usage{
remove_flagged(x, strict = TRUE)
}
\arguments{
\item{x}{Object of class \code{ghcn_daily}. See \code{\link[=daily]{daily()}} for details.}

\item{strict}{Logical, if to remove also \code{looser} flags.}
}
\value{
\code{x} without flagged records.
}
\description{
Remove Flagged Recrods
}
\examples{
remove_flagged(CA003076680)
}
