% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{remove_flagged}
\alias{remove_flagged}
\title{Remove Flagged Recrods}
\usage{
remove_flagged(x, strict = FALSE)
}
\arguments{
\item{x}{Object of class \code{daily}.}

\item{strict}{Logical, if to remove also \code{looser} flags.}
}
\value{
A tibble with the stations within the \code{roi}.
}
\description{
Remove Flagged Recrods
}
\details{
dates should be given in \code{YYYY-mm-dd} format.
}
\examples{
remove_flagged(CA003076680)
}
