% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{Weather Stations from GHCN}
\format{
\subsection{\code{stations}}{

A table 756,939 rows and 6 columns:
\describe{
\item{station}{Station name}
\item{longitude}{Longitude}
\item{latitude}{Latitude}
\item{variable}{Weather variable}
\item{startYear}{First year of recording}
\item{lastYear}{Last year of recording}
}
}
}
\source{
\url{https://www.ncei.noaa.gov/pub/data/ghcn/daily/ghcnd-inventory.txt}
}
\usage{
stations

stations(roi, show = FALSE)
}
\arguments{
\item{roi}{SpatVector defining the region of interest.}

\item{show}{Logical, show the stations on a map?}
}
\value{
A tibble with the stations within the \code{roi}.
}
\description{
Weather Stations from GHCN

Get Stations for Area
}
\keyword{datasets}
