% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R, R/utils.R
\name{stations}
\alias{stations}
\title{Get Stations for Area}
\usage{
stations(filename, variables = c("tmin", "tmax", "prcp"))

stations(filename, variables = c("tmin", "tmax", "prcp"))
}
\arguments{
\item{filename}{Character of the filename of the inventory, if already downloaded.}

\item{variables}{Character, vector of the variables to include.}

\item{roi}{SpatVector defining the region of interest.}

\item{show}{Logical, show the stations on a map?}
}
\value{
A tibble with the stations within the \code{roi}.

A tibble with the station inventory.
}
\description{
Get Stations for Area

Get GHCNd Inventory
}
\details{
If \emph{filename} is not provided, this will download the inventory from
<"https://www.ncei.noaa.gov/pub/data/ghcn/daily/ghcnd-inventory.txt">.
In alternative, you can download the invetory yourself and load it (see examples).
}
\examples{
\dontrun{
dest <- tempfile()
download.file("https://www.ncei.noaa.gov/pub/data/ghcn/daily/ghcnd-inventory.txt", dest)
s <- stations(dest)
}
}
