% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations}
\alias{stations}
\title{Get Stations for Area}
\usage{
stations(roi, show = FALSE)
}
\arguments{
\item{roi}{SpatVector defining the region of interest.}

\item{show}{Logical, show the stations on a map?}
}
\value{
A tibble with the stations within the \code{roi}.
}
\description{
Get Stations for Area
}
